test/reactor_ut: don't assert number of events
Refactor this part of the unit tests to make it a bit
easier to maintain as the dynamic scheduler itself is
modified.

For example, depending on the simulated thread loads,
we may need to pass extra events to cores for
purposes of setting interrupt mode.  The important
thing to test here isn't how many events it takes to
do that, but what is the end result.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Iad2e861cfa0bfd16c853332650e3ab3a9727f490

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/9624
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>